temp fix

+++ src/main/java/rest/order/reservation/Model/Menu.java
@@ -26,7 +26,7 @@ public class Menu {
     public Menu() {
     }
 
-    public Menu(Long mid, String name, String intro, int price, MenuType type) {
+    public Menu(Long mid, String name, String intro, Long price, MenuType type) {
         this.mid = mid;
         this.name = name;
         this.intro = intro;
@@ -34,7 +34,7 @@ public class Menu {
         this.type = type;
     }
 
-    public Menu(String name, String intro, int price, MenuType type) {
+    public Menu(String name, String intro, Long price, MenuType type) {
         this.name = name;
         this.intro = intro;
         this.price = price;
@@ -43,7 +43,7 @@ public class Menu {
 
     // 메뉴 수정
     // set 사용 x
-    public void changeMenuInfo(String name, String intro, int price, MenuType type) {
+    public void changeMenuInfo(String name, String intro, Long price, MenuType type) {
         this.name = name;
         this.intro = intro;
         this.price = price;
@@ -74,11 +74,11 @@ public class Menu {
         this.intro = intro;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
